--- src/filter/env_filter/directive.rs
@@ -1,34 +1,0 @@
-// Copyright 2024 FastLabs Developers
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use log::Level;
-use log::LevelFilter;
-
-#[derive(Debug)]
-pub(crate) struct Directive {
-    pub(crate) name: Option<String>,
-    pub(crate) level: LevelFilter,
-}
-
-// Check whether a level and target are enabled by the set of directives.
-pub(crate) fn enabled(directives: &[Directive], level: Level, target: &str) -> bool {
-    // Search for the longest match, the vector is assumed to be pre-sorted.
-    for directive in directives.iter().rev() {
-        match directive.name {
-            Some(ref name) if !target.starts_with(&**name) => {}
-            Some(..) | None => return level <= directive.level,
-        }
-    }
-    false
-}
